+++ examples/demo3/test_1_0.v
@@ -36,15 +36,13 @@ End Theory.
 
 (* Instance *)
 
+HB.instance
 Definition Z_mulmonoid_axioms :=
   MulMonoid_of_Type.Build Z 1%Z Z.mul Z.mul_assoc Z.mul_1_l Z.mul_1_r.
 
-HB.instance Z Z_mulmonoid_axioms.
-
+HB.instance
 Definition Z_ring_axioms :=
   Ring_of_MulMonoid.Build Z 0%Z Z.add
     Z.add_assoc Z.add_0_l Z.add_0_r
     Z.opp Z.add_comm Z.add_opp_diag_l
     Z.mul_add_distr_r Z.mul_add_distr_l.
-
-HB.instance Z Z_ring_axioms.